Added nice scanner and progress bars

+++ UI/LogData.cs
@@ -11,25 +11,25 @@ namespace SteerLoggerUser
         public List<string> rawheaders = new List<string>();
         public List<string> convheaders = new List<string>();
         public List<DateTime> timestamp = new List<DateTime>();
-        public List<decimal> time = new List<decimal>();
-        public List<List<decimal>> rawData = new List<List<decimal>>();
-        public List<List<decimal>> convData = new List<List<decimal>>();
+        public List<double> time = new List<double>();
+        public List<List<double>> rawData = new List<List<double>>();
+        public List<List<double>> convData = new List<List<double>>();
 
         // Initialises rawData and convData lists by populating them with correct number of empty lists
         public void InitRawConv(int pinNum)
         {
             for (int i = 0; i < pinNum; i++)
             {
-                this.rawData.Add(new List<decimal>());
-                this.convData.Add(new List<decimal>());
+                this.rawData.Add(new List<double>());
+                this.convData.Add(new List<double>());
             }
         }
         
         // Adds a list of values to the corresponding rawData columns
-        public void AddRawData(List<decimal> values)
+        public void AddRawData(List<double> values)
         {
             int i = 0;
-            foreach (decimal value in values)
+            foreach (double value in values)
             {
                 this.rawData[i].Add(value);
                 i += 1;
@@ -37,10 +37,10 @@ namespace SteerLoggerUser
         }
 
         // Adds a list of values to the corresponding convData columns
-        public void AddConvData(List<decimal> values)
+        public void AddConvData(List<double> values)
         {
             int i = 0;
-            foreach (decimal value in values)
+            foreach (double value in values)
             {
                 this.convData[i].Add(value);
                 i += 1;
