+++ UI/LogProc.cs
@@ -10,8 +10,8 @@ namespace SteerLoggerUser
     {
         public List<string> procheaders = new List<string>();
         public List<DateTime> timestamp = new List<DateTime>();
-        public List<decimal> time = new List<decimal>();
-        public List<List<decimal>> procData = new List<List<decimal>>();
+        public List<double> time = new List<double>();
+        public List<List<double>> procData = new List<List<double>>();
 
 
         // Sets the LogProc values to match the values for converted data of LogData
@@ -28,15 +28,15 @@ namespace SteerLoggerUser
         {
             for (int i = 0; i < pinNum; i++)
             {
-                this.procData.Add(new List<decimal>());
+                this.procData.Add(new List<double>());
             }
         }
         
         // Adds a list of values to the corresponding procData columns
-        public void AddProcData(List<decimal> values)
+        public void AddProcData(List<double> values)
         {
             int i = 0;
-            foreach (decimal value in values)
+            foreach (double value in values)
             {
                 this.procData[i].Add(value);
                 i += 1;
